feat: extracts baseSamlConfig andgetSamlOptionsForRequest
